+++ src/hooks.server.ts
@@ -1,11 +1,12 @@
 import { createServerClient } from '@supabase/ssr';
 import { env } from '$env/dynamic/public';
 import type { Handle } from '@sveltejs/kit';
+import { PUBLIC_SUPABASE_PUBLISHABLE_DEFAULT_KEY, PUBLIC_SUPABASE_URL } from '$env/static/public';
 
 export const handle: Handle = async ({ event, resolve }) => {
 	event.locals.supabase = createServerClient(
-		env.PUBLIC_SUPABASE_URL,
-		env.PUBLIC_SUPABASE_PUBLISHABLE_DEFAULT_KEY,
+		PUBLIC_SUPABASE_URL,
+		PUBLIC_SUPABASE_PUBLISHABLE_DEFAULT_KEY,
 		{
 			cookies: {
 				getAll: () => event.cookies.getAll(),